Move store to credentials
