+++ include/credentials.h
@@ -1,4 +1,9 @@
 #include <stdlib.h>
+#include <stdio.h>
+#include "password.h"
+
+char storeCredentials(Credentials_t);
+
 
 typedef struct Credentials_t {
     char *label;
