--- include/store.h
@@ -1,5 +1,0 @@
-#include <stdio.h>
-#include "password.h"
-#include "credentials.h"
-
-char storeCredentials(Credentials_t);
